解决人物移动bug

+++ Dungeon_Game/src/hero.h
@@ -5,7 +5,7 @@
 
 //∂®“ÂHERO
 struct HERO {
-	float position_x, position_y;
+	int position_x, position_y;
 	int HP;
 	int attack;
 	int defence;
